+++ tool.py
@@ -1,12 +1,8 @@
 #!/usr/bin/env python3
-import os
 import json
 import re
 import shlex
-import struct
-import subprocess
 import sys
-import threading
 import time
 import tkinter as tk
 from configparser import ConfigParser
@@ -14,7 +10,7 @@ import extra
 from extra import *
 import contextpatch
 import utils
-
+from utils import CallZ, formats
 if os.name == 'nt':
     import windnd
 import zipfile
@@ -53,18 +49,7 @@ theme = StringVar()
 language = StringVar()
 var1 = IntVar()
 car = IntVar()
-formats = ([b'PK', "zip"], [b'OPPOENCRYPT!', "ozip"], [b'7z', "7z"], [b'\x53\xef', 'ext', 1080],
-           [b'\x3a\xff\x26\xed', "sparse"], [b'\xe2\xe1\xf5\xe0', "erofs", 1024], [b"CrAU", "payload"],
-           [b"AVB0", "vbmeta"], [b'\xd7\xb7\xab\x1e', "dtbo"],
-           [b'\xd0\x0d\xfe\xed', "dtb"], [b"MZ", "exe"], [b".ELF", 'elf'],
-           [b"ANDROID!", "boot"], [b"VNDRBOOT", "vendor_boot"],
-           [b'AVBf', "avb_foot"], [b'BZh', "bzip2"],
-           [b'CHROMEOS', 'chrome'], [b'\x1f\x8b', "gzip"],
-           [b'\x1f\x9e', "gzip"], [b'\x02\x21\x4c\x18', "lz4_legacy"],
-           [b'\x03\x21\x4c\x18', 'lz4'], [b'\x04\x22\x4d\x18', 'lz4'],
-           [b'\x1f\x8b\x08\x00\x00\x00\x00\x00\x02\x03', "zopfli"], [b'\xfd7zXZ', 'xz'],
-           [b']\x00\x00\x00\x04\xff\xff\xff\xff\xff\xff\xff\xff', 'lzma'], [b'\x02!L\x18', 'lz4_lg'],
-           [b'\x89PNG', 'png'], [b"LOGO!!!!", 'logo'])
+
 
 
 class ModuleError(Exception):
@@ -423,8 +408,7 @@ def getframe(title):
 
 
 # 子进程运行 防卡死
-def CallZ(func, *args):
-    threading.Thread(target=func, args=args, daemon=True).start()
+
 
 
 def subp(com: int = 1, title: str = lang.text18, master: any = None):
