+++ utils.py
@@ -4,7 +4,7 @@ import struct
 from os.path import exists
 import sys, os, errno, tempfile
 import common, blockimgdiff, sparse_img
-
+from threading import Thread
 
 # -----
 # ====================================================
@@ -13,6 +13,20 @@ import common, blockimgdiff, sparse_img
 #          DATE: 2018-05-25 12:19:12 CEST
 # ====================================================
 # -----
+formats = ([b'PK', "zip"], [b'OPPOENCRYPT!', "ozip"], [b'7z', "7z"], [b'\x53\xef', 'ext', 1080],
+           [b'\x3a\xff\x26\xed', "sparse"], [b'\xe2\xe1\xf5\xe0', "erofs", 1024], [b"CrAU", "payload"],
+           [b"AVB0", "vbmeta"], [b'\xd7\xb7\xab\x1e', "dtbo"],
+           [b'\xd0\x0d\xfe\xed', "dtb"], [b"MZ", "exe"], [b".ELF", 'elf'],
+           [b"ANDROID!", "boot"], [b"VNDRBOOT", "vendor_boot"],
+           [b'AVBf', "avb_foot"], [b'BZh', "bzip2"],
+           [b'CHROMEOS', 'chrome'], [b'\x1f\x8b', "gzip"],
+           [b'\x1f\x9e', "gzip"], [b'\x02\x21\x4c\x18', "lz4_legacy"],
+           [b'\x03\x21\x4c\x18', 'lz4'], [b'\x04\x22\x4d\x18', 'lz4'],
+           [b'\x1f\x8b\x08\x00\x00\x00\x00\x00\x02\x03', "zopfli"], [b'\xfd7zXZ', 'xz'],
+           [b']\x00\x00\x00\x04\xff\xff\xff\xff\xff\xff\xff\xff', 'lzma'], [b'\x02!L\x18', 'lz4_lg'],
+           [b'\x89PNG', 'png'], [b"LOGO!!!!", 'logo'])
+
+
 def qc(file_) -> None:
     if not exists(file_):
         return
@@ -26,6 +40,10 @@ def qc(file_) -> None:
         f.truncate()
         f.writelines(new_data)
     del data, new_data
+
+
+def CallZ(func, *args):
+    Thread(target=func, args=args, daemon=True).start()
 
 
 def img2sdat(input_image, out_dir='.', version=None, prefix='system'):
